+++ test/test_optimisation.js
@@ -602,6 +602,7 @@ QUnit.test('Continuous quadratic knapsack problem solver - Breakpoint searching 
 	}
 });
 
+
 QUnit.test('Linear programming solver - Primal Dual Hybrid Gradient', function(assert) {    
   // Reference:  Vanderbei, Robert J, Linear Programming Foundations and Extensions 4th edition
   // Problem, p. 11