+++ changelog.md
@@ -1,4 +1,5 @@
 ### 0.0.6 - XX/YY/2018
+- Updated the Maximin portfolio method output signature to allow outputting only the portfolio weights
 
 ### 0.0.5 - 02/04/2018
 - Misc. Matrix/Vector operations additions/refactor (in particular, toDoubleArray method replaced by toRowArray)